이미지, 사이트유저 엔티티 정리 및 이미지 파일 업로드 로직 수정

+++ src/main/java/com/project/echoproject/service/MypageServiceImpl.java
@@ -11,7 +11,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
-import java.util.Optional;
 
 @Service  // 이 클래스가 서비스 계층의 빈으로 등록되도록 표시합니다.
 public class MypageServiceImpl implements MypageService {
@@ -35,11 +34,11 @@ public class MypageServiceImpl implements MypageService {
     }
 
     // 주어진 사용자 ID에 해당하는 사용자의 정보를 업데이트합니다.
+    @Override
     public void updateUser(String userId, SiteUser updatedUser, MultipartFile file) throws IOException {
         SiteUser updateInfo = userRepository.findByUserId(userId)
                 .orElseThrow(() -> new IllegalArgumentException("User not found"));
 
-        // 사용자 정보를 업데이트합니다.
         updateInfo.setUserName(updatedUser.getUserName());
         updateInfo.setNickName(updatedUser.getNickName());
         updateInfo.setEmail(updatedUser.getEmail());
@@ -48,20 +47,14 @@ public class MypageServiceImpl implements MypageService {
         updateInfo.setStreetaddr(updatedUser.getStreetaddr());
         updateInfo.setDetailaddr(updatedUser.getDetailaddr());
 
-        // 파일이 비어 있지 않은 경우 프로필 이미지를 업데이트합니다.
         if (file != null && !file.isEmpty()) {
-            // 기존 프로필 이미지가 있다면 삭제합니다.
-            if (updateInfo.getProfileImage() != null) {
+            Image newImage = imageService.saveImage(file);
+            if (updateInfo.getProfileImage() != null && !updateInfo.getProfileImage().isDefaultImage()) {
                 imageService.deleteImage(updateInfo.getProfileImage());
             }
-
-            // 새 이미지를 저장합니다.
-            Image newImage = imageService.saveImage(file);
             updateInfo.setProfileImage(newImage);
-            newImage.setSiteUser(updateInfo);  // 이미지 엔티티와 사용자 엔티티를 연관시킵니다.
         }
 
-        // 업데이트된 사용자 정보를 저장합니다.
         userRepository.save(updateInfo);
     }
 
@@ -80,7 +73,7 @@ public class MypageServiceImpl implements MypageService {
 
         if (updatedUser.getProfileImage() != null && !updatedUser.getProfileImage().isEmpty()) {
             Image newImage = imageService.saveImage(updatedUser.getProfileImage());
-            if (user.getProfileImage() != null) {
+            if (user.getProfileImage() != null && !user.getProfileImage().isDefaultImage()) {
                 imageService.deleteImage(user.getProfileImage());
             }
             user.setProfileImage(newImage);
